Module Pgx_async_test

Testing library for code that uses postgres

val set_to_default_db : unit -> unit

type 'a new_db_callback = Pgx_async.t -> db_name:string -> 'a Async_kernel.Deferred.t

val with_temp_db : 'a new_db_callback -> 'a Async_kernel.Deferred.t

with_temp_db f creates a temporary database and executes f with a database handle to this db and the name of the db. Once f executes or raises, the temp database will be deleted.